refactored: created encoder library
