+++ Encoder/IEncoder.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Encoder
+{
+    public interface IEncoder
+    {
+        void Encode(byte[] source);
+    }
+}
