+++ EventsDelegates.sln
@@ -3,9 +3,11 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "External", "External\External.csproj", "{D78AB70E-1FED-4D66-B2B0-FA0B7F78C1C2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "External", "External\External.csproj", "{D78AB70E-1FED-4D66-B2B0-FA0B7F78C1C2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Demo", "Demo\Demo.csproj", "{4CC3F7F8-319E-4A68-A5F6-C93AFEC6FCF4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo", "Demo\Demo.csproj", "{4CC3F7F8-319E-4A68-A5F6-C93AFEC6FCF4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Encoder", "Encoder\Encoder.csproj", "{B8B68068-4971-4D09-8B37-44EB52D7769A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{4CC3F7F8-319E-4A68-A5F6-C93AFEC6FCF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4CC3F7F8-319E-4A68-A5F6-C93AFEC6FCF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4CC3F7F8-319E-4A68-A5F6-C93AFEC6FCF4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B8B68068-4971-4D09-8B37-44EB52D7769A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B8B68068-4971-4D09-8B37-44EB52D7769A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B8B68068-4971-4D09-8B37-44EB52D7769A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B8B68068-4971-4D09-8B37-44EB52D7769A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
