--- External/Encoder/IEncoder.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace External.Encoder
-{
-    public interface IEncoder
-    {
-        void Encode(byte[] source);
-
-        void Notify(string message);
-
-        event EventHandler<EncodedEventArgs> Encoded;
-    }
-}
